Changes: 1.
Updated some of the the user input on menu selection to use:
getline(cin, mychoicestr);
stringstream(mychoicestr) << choice;
instead of:
cin>>choice;
This change will prevent the program from being stuck in an infnite loop (as it did before this change) if the user enters in something besides an integer. However, this can not be used for when the user has to enter a number with a decimal, so I clarified some of the user input to specify this (I'm assuming that the user does not know what a 'float' is).

2. Removed 'dummy' string cin's as they prevented the user from entering the correct name. For example, when I tried to add the hero Shoe it would enter the name n because I had to hit the enter key twice for it to accecpt my input. Upon further corrections, I have noticed that the getline/cin combo that is used in this program is causing some erratic errors. I tried to fix as many as I possibly could, but using another method of user input may be a better alternative.